requirements.txt
Django==1.8.6
django-extensions==1.5.9
djangorestframework==3.3.1
